optimized DB accesses during history download/parse

+++ src/net/bicou/redmine/data/sqlite/TrackersDbAdapter.java
@@ -84,7 +84,7 @@ public class TrackersDbAdapter extends DbAdapter {
 		return list;
 	}
 
-	public Tracker select(final Server server, final int id) {
+	public Tracker select(final Server server, final long id) {
 		if (server == null) {
 			return null;
 		}
